# web/lib/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StoreOps.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/orders": {
      "get": {
        "tags": [
          "AdminOrders"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminOrderSummaryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminOrderSummaryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminOrderSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders/{orderNumber}": {
      "get": {
        "tags": [
          "AdminOrders"
        ],
        "parameters": [
          {
            "name": "orderNumber",
            "in": "path",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminOrders"
        ],
        "parameters": [
          {
            "name": "orderNumber",
            "in": "path",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/orders/bulk-status": {
      "post": {
        "tags": [
          "AdminOrders"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkOrderStatusRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkOrderStatusRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BulkOrderStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultOfstring"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultOfstring"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultOfstring"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/account/orders/{orderNumber}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "orderNumber",
            "in": "path",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/customers": {
      "get": {
        "tags": [
          "AdminCustomers"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AccountRole"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminCustomerSummaryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminCustomerSummaryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminCustomerSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/customers/{id}": {
      "get": {
        "tags": [
          "AdminCustomers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/customers/{id}/disable": {
      "post": {
        "tags": [
          "AdminCustomers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/customers/{id}/enable": {
      "post": {
        "tags": [
          "AdminCustomers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/checkout": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCheckoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCheckoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkout/result": {
      "get": {
        "tags": [
          "Checkout"
        ],
        "parameters": [
          {
            "name": "paymentIntentId",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResultResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResultResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResultResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountRole": {
        "enum": [
          "customer",
          "admin",
          null
        ]
      },
...
      "AdminCustomerResponse": {
        "required": [
          "id",
          "username",
          "email",
          "joinedAtUtc",
          "isAdmin",
          "isDisabled",
          "orderCount",
          "spentCents",
          "addresses",
          "recentOrders"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "joinedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "isDisabled": {
            "type": "boolean"
          },
          "orderCount": {
            "type": "integer",
            "format": "int32"
          },
          "spentCents": {
            "type": "integer",
            "format": "int32"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressResponse"
            }
          },
          "recentOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderSummaryResponse"
            }
          }
        }
      },
...
      "AdminOrderResponse": {
        "required": [
          "orderNumber",
          "status",
          "nextStatuses",
          "placedAtUtc",
          "customerId",
          "customerName",
          "customerEmail",
          "lines",
          "subtotalCents",
          "shippingCents",
          "taxCents",
          "totalCents",
          "shipTo",
          "carrier",
          "trackingNumber",
          "trackingUrl",
          "estimatedDeliveryDate",
          "timeline",
          "stripePaymentIntentId",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "nextStatuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          "placedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "customerId": {
            "type": "integer",
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineResponse"
            }
          },
          "subtotalCents": {
            "type": "integer",
            "format": "int32"
          },
          "shippingCents": {
            "type": "integer",
            "format": "int32"
          },
          "taxCents": {
            "type": "integer",
            "format": "int32"
          },
          "totalCents": {
            "type": "integer",
            "format": "int32"
          },
          "shipTo": {
            "$ref": "#/components/schemas/PostalAddress"
          },
          "carrier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Carrier"
              }
            ]
          },
          "trackingNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "trackingUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "estimatedDeliveryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderStepResponse"
            }
          },
          "stripePaymentIntentId": {
            "type": "string"
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "AdminOrderSummaryResponse": {
        "required": [
          "orderNumber",
          "status",
          "placedAtUtc",
          "customerName",
          "customerEmail",
          "itemCount",
          "totalCents"
        ],
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "placedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "customerName": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "itemCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalCents": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "BulkOrderStatusRequest": {
        "required": [
          "orderNumbers"
        ],
        "type": "object",
        "properties": {
          "orderNumbers": {
            "maxItems": 100,
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "emailCustomer": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "confirmRefund": {
            "type": "boolean"
          }
        }
      },
...
      "BulkResultOfstring": {
        "required": [
          "succeeded",
          "failed"
        ],
        "type": "object",
        "properties": {
          "succeeded": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "failed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkFailureOfstring"
            }
          }
        }
      },
...
      "CheckoutResponse": {
        "required": [
          "checkoutId",
          "clientSecret",
          "lines",
          "shipTo",
          "subtotalCents",
          "shippingCents",
          "taxCents",
          "totalCents"
        ],
        "type": "object",
        "properties": {
          "checkoutId": {
            "type": "integer",
            "format": "int32"
          },
          "clientSecret": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutLineResponse"
            }
          },
          "shipTo": {
            "$ref": "#/components/schemas/PostalAddress"
          },
          "subtotalCents": {
            "type": "integer",
            "format": "int32"
          },
          "shippingCents": {
            "type": "integer",
            "format": "int32"
          },
          "taxCents": {
            "type": "integer",
            "format": "int32"
          },
          "totalCents": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CheckoutResultResponse": {
        "required": [
          "result",
          "orderNumber"
        ],
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/PaymentResult"
          },
          "orderNumber": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "OrderResponse": {
        "required": [
          "orderNumber",
          "status",
          "placedAtUtc",
          "lines",
          "subtotalCents",
          "shippingCents",
          "taxCents",
          "totalCents",
          "shipTo",
          "carrier",
          "trackingNumber",
          "trackingUrl",
          "estimatedDeliveryDate",
          "timeline"
        ],
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "placedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineResponse"
            }
          },
          "subtotalCents": {
            "type": "integer",
            "format": "int32"
          },
          "shippingCents": {
            "type": "integer",
            "format": "int32"
          },
          "taxCents": {
            "type": "integer",
            "format": "int32"
          },
          "totalCents": {
            "type": "integer",
            "format": "int32"
          },
          "shipTo": {
            "$ref": "#/components/schemas/PostalAddress"
          },
          "carrier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Carrier"
              }
            ]
          },
          "trackingNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "trackingUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "estimatedDeliveryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStepResponse"
            }
          }
        }
      },
      "OrderStatus": {
        "enum": [
          "pending",
          "shipped",
          "delivered",
          "cancelled",
          "refunded"
        ]
      },
...
      "PagedOfAdminCustomerSummaryResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCustomerSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PagedOfAdminOrderSummaryResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StartCheckoutRequest": {
        "type": "object",
        "properties": {
          "addressId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateOrderRequest": {
        "required": [
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "carrier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Carrier"
              }
            ]
          },
          "trackingNumber": {
            "maxLength": 100,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "estimatedDeliveryDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "note": {
            "maxLength": 300,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "emailCustomer": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "confirmRefund": {
            "type": "boolean"
          },
          "rowVersion": {
            "maxLength": 8,
            "minLength": 8,
            "type": "string",
            "format": "byte"
          }
        }
      },
...
    }
  },
...
}